+++ src/Randomizer/SystemRandomizer.php
@@ -4,11 +4,21 @@ declare(strict_types=1);
 
 namespace Rasuvaeff\Retry\Randomizer;
 
+use Random\IntervalBoundary;
+use Random\Randomizer;
+
 /**
  * @api
  */
 final readonly class SystemRandomizer implements RandomizerInterface
 {
+    private Randomizer $randomizer;
+
+    public function __construct()
+    {
+        $this->randomizer = new Randomizer();
+    }
+
     #[\Override]
     public function float(float $min, float $max): float
     {
@@ -16,6 +26,10 @@ final readonly class SystemRandomizer implements RandomizerInterface
             throw new \InvalidArgumentException('Minimum must be less than or equal to maximum');
         }
 
-        return $min + (lcg_value() * ($max - $min));
+        if ($min === $max) {
+            return $min;
+        }
+
+        return $this->randomizer->getFloat($min, $max, IntervalBoundary::ClosedClosed);
     }
 }